refactor : 공모전 엔티티 디렉토리명 변경
